:recycle: refactor: 디렉토리 경로 수정
feature 통합
아이콘 shared 이동
