+++ src/pages/my/components/modal/ConfirmationModal.tsx
@@ -2,7 +2,7 @@ import { useRef, useEffect } from 'react';
 import { useNavigate } from 'react-router';
 import styled from 'styled-components';
 import ReactDOM from 'react-dom';
-import { deletePostMutation } from '@pages/my/hooks/mutate/useDeletePost';
+import { deletePostMutation } from '@my/features/hooks/mutate/useDeletePost';
 import { usePostIdStore } from '@pages/my/features/store/usePostIdStore';
 import useIsLoggedInStore from '@shared/store/useIsLoggedIn';
 import { apiInstance } from '@shared/apis/instance';
