+++ src/pages/my/components/modal/EditUserInfoModal.tsx
@@ -5,8 +5,8 @@ import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { IcLeftArrow } from '@shared/assets/icon/ic-left-arrow';
 import { ModalProps } from '@shared/types/my/modalProps';
-import Human from '@onboarding/icons/ic-human.svg?react';
-import Edit from '@onboarding/icons/ic-edit.svg?react';
+import Human from '@icon/ic-human.svg?react';
+import Edit from '@icon/ic-edit.svg?react';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { UserInfoSchema } from '@pages/my/features/schema/MyUserInfoSchema';
 import { UserDataInput } from '@my/components/UserDataInput';
