+++ src/pages/onboarding/components/HeightModal.tsx
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
 import { useState, useRef } from 'react';
 import { UseFormWatch, UseFormSetValue } from 'react-hook-form';
-import { UserInfoSchemaType } from '@onboarding/schema';
+import { UserInfoSchemaType } from 'pages/onboarding/feature/schema';
 import Dialog from '@pages/onboarding/components/HeightDialog';
-import usePositionedDialog from '@onboarding/hooks/usePositionedDialog';
-import useOutsideClick from '@onboarding/hooks/useOutsideClick';
+import usePositionedDialog from '@onboarding/feature/hooks/usePositionedDialog';
+import useOutsideClick from '@onboarding/feature/hooks/useOutsideClick';
 
 const BirthdayModal = ({
   watch,
