+++ src/pages/onboarding/feature/hooks/useOutsideClick.ts
