+++ src/pages/onboarding/types/index.ts
@@ -1,5 +1,5 @@
 import { UseFormRegister, UseFormWatch, UseFormSetValue, UseFormGetValues } from 'react-hook-form';
-import { UserInfoSchemaType } from '@onboarding/schema';
+import { UserInfoSchemaType } from 'pages/onboarding/feature/schema';
 
 export type RegisterType = UseFormRegister<UserInfoSchemaType>;
