+++ src/pages/post/components/CreatePost.tsx
@@ -8,7 +8,7 @@ import { CreateNewPostModal } from '@pages/post/components/modal/CreateNewPostMo
 import { SelectPhotoBottomSheetModal } from '@pages/post/components/modal/SelectPhotoBottomSheetModal';
 import { createPresignedUrl } from '@pages/post/apis/createPresignedUrl';
 import { presignedUrlProps } from '@pages/post/apis/createPresignedUrl';
-import { useControlModal } from '@pages/my/hooks/useControlModal';
+import { useControlModal } from '@my/features/hooks/useControlModal';
 export const CreatePost = ({ isOpened }: { isOpened: boolean }) => {
   const { modalState, setModalState } = useControlModal();
   const [opened, setIsOpened] = useState<boolean>(isOpened);
